Reimplement with constructors, test with 3 htslib versions
