--- test_CpG.bedGraph
@@ -1,1 +1,0 @@
-track type="bedGraph" description="test CpG methylation levels"
